做了大量修改
1. 修改了readme，修正了一些细节问题以及使用pip时换tuna源
2. 修改了文件组织结构
3. 修改.py成可执行文件
4. 修改了requirements.txt，原来的有一些问题
