--- controller.py
@@ -1,4 +1,0 @@
-from PyQt5.QtCore import QObject, pyqtSignal
-class Controller(QObject):
-    clip_changed = pyqtSignal(str)
-con = Controller()
