+++ requirements.txt
@@ -1,5 +1,4 @@
-pyperclip==1.7.0
-PyQt5==5.13.1
-PyQt5-sip==12.7.0
-PyQtWebEngine==5.13.2
-requests==2.22.0
+pyperclip
+pyqt5
+requests
+pyqtWebEngine
